DTO usage and swapped decimal DB values out for integers + new demo accounts

+++ database/migrations/2026_01_16_101024_add_balance_to_users_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->decimal('balance', 15, 2)->default(0.00)->after('role');
+            $table->bigInteger('balance')->default(0)->after('role');
         });
     }
